fix: fixed issue #1: Project fails to run initially due to incorrect file structure and extensions
