+++ Frontend/src/imports.js
@@ -1,6 +1,6 @@
 import Home from "./views/pages/Home"
-import Cart  from "./views/components/cart"
-import Itemframes from "./views/components/itemframes"
+import Cart  from "./views/components/Cart"
+import Itemframes from "./views/components/Itemframes"
 import Customers from "./views/components/Customers"
 import Orders from "./views/pages/Orders"
